fix regressions #34 & #35
