+++ src/Frontend/src/js/webhooks.js
@@ -1,16 +1,7 @@
 ﻿"use strict";
 
 import { HubConnectionBuilder } from "@microsoft/signalr";
-import { Tooltip } from "bootstrap";
 
-window.Dropdown = $.fn.dropdown;
-import 'bootstrap-select';
-
-import $ from 'jquery';
-window.jQuery = $;
-window.$ = $;
-
-import 'bootstrap-select/dist/css/bootstrap-select.css';
 import '../css/webhooks.scss';
 
 let nodes = new Array();
@@ -21,9 +12,6 @@ $.fn.selectpicker.Constructor.BootstrapVersion = '4';
 
 var connection = new HubConnectionBuilder().withUrl("/webhookHub").build();
 connection.serverTimeoutInMilliseconds = 120000;
-
-const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
-const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new Tooltip(tooltipTriggerEl))
 
 $("#enable-verification").click(function () {
     if ($(this).is(':checked')) {
